+++ Sources/MapboxMaps/Gestures/GestureHandlers/DoubleTapToZoomInGestureHandler.swift
@@ -21,7 +21,7 @@ internal final class DoubleTapToZoomInGestureHandler: GestureHandler {
         case .recognized:
             cameraAnimationsManager.cancelAnimations()
             delegate?.gestureBegan(for: .doubleTapToZoomIn)
-            delegate?.gestureEnded(for: .doubleTapToZoomIn, willDecelerate: true)
+            delegate?.gestureEnded(for: .doubleTapToZoomIn, willAnimate: true)
             cameraAnimationsManager.ease(
                 to: CameraOptions(zoom: mapboxMap.cameraState.zoom + 1),
                 duration: 0.3,